♻️ Cleanup the lonely `index.ts`
